pyproject.toml
[tool.poetry]
name = "hw-24-1-2-"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "hw_24"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
